Separate integration test suite
