+++ webpack.config.js
@@ -31,6 +31,7 @@ const config = () => {
         entry: {
             extension: "./src/extension.ts",
             "testBundle.test": glob.sync("./src/test/**/*.test.ts"),
+            "integration.spec": glob.sync("./src/test-integration/**/*.spec.ts"),
         },
         output: {
             path: path.resolve(__dirname, "dist"),
